% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainflowprobes-package.R
\docType{package}
\name{brainflowprobes-package}
\alias{brainflowprobes}
\alias{brainflowprobes-package}
\title{brainflowprobes: Plots and annotation for choosing BrainFlow target probe sequence}
\description{
Use these functions to characterize genomic regions for
  BrainFlow target probe design.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LieberInstitute/brainflowprobes}
  \item Report bugs at \url{https://github.com/LieberInstitute/brainflowprobes/issues}
}

}
\author{
\strong{Maintainer}: Amanda Price \email{amanda.joy.price@gmail.com} (\href{https://orcid.org/0000-0001-7352-3732}{ORCID})

Other contributors:
\itemize{
  \item Leonardo Collado-Torres \email{lcolladotor@gmail.com} (\href{https://orcid.org/0000-0003-2140-308X}{ORCID}) [contributor]
}

}
\keyword{internal}
