% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-data.R
\docType{data}
\name{four_panels_example_cov}
\alias{four_panels_example_cov}
\title{Example base-pair region coverage data}
\format{A list with the base-pair coverage output from
\link{brainflowprobes_cov} with the example region used throughout the
package documentation ('chr20:10286777-10288069:+').
\describe{
\item{Sep }{ base-pair coverage region data.frame list for pd$Sep,}
\item{Deg }{ base-pair coverage region data.frame list for pd$Deg,}
\item{Cell }{ base-pair coverage region data.frame list for pd$Cell,}
\item{Sort }{ base-pair coverage region data.frame list for pd$Sort.}
}}
\description{
A list of base-pair region coverage data.frame lists used for exemplifying
the package functionality. This is the data extracted for the example region
\code{chr20:10286777-10288069:+} by looping through the phenotype tables
stored in \code{pd} and using the
\link[derfinder:getRegionCoverage]{getRegionCoverage} function. This can be
reproduced using the \link{brainflowprobes_cov} function in this package.
}
\examples{
if(FALSE) {
    ## Takes about 10 minutes to run!
    four_panels_example_cov <- brainflowprobes_cov(
        'chr20:10286777-10288069:+'
    )
}
}
\seealso{
\link{four_panels} \link{plot_coverage}
}
\keyword{datasets}
