% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-data.R
\docType{data}
\name{genes}
\alias{genes}
\title{GRanges object with the annotated genes for hg19}
\format{A GRanges object with the following annotation columns.
\describe{
\item{CSS }{ the coding region start (NA for non-coding genes),}
\item{CSE }{ the coding region end (NA for non-coding genes),}
\item{Tx }{ the Gencode Gene ID,}
\item{Geneid }{ the Ensembl Gene ID,}
\item{Gene }{ the Gene symbol,}
\item{Refseq }{ the RefSeq gene ID,}
\item{Nexons }{ the number of exons,}
\item{Exons }{ the exon coordinates.}
}
See \link[bumphunter:annotateTranscripts]{annotateTranscripts} for more
information.}
\description{
A GRanges object with the annotated genes for hg19 using
\code{bumphunter::annotateTranscripts} on the
Gencode v31 annotation lifted over to hg19 grouped by gene
instead of by transcript.
}
\seealso{
\link{four_panels}
\url{https://github.com/LieberInstitute/brainflowprobes/blob/master/data-raw/create_sysdata.R}
\url{https://www.gencodegenes.org/human/release_31lift37.html}
}
\keyword{datasets}
