% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-data.R
\docType{data}
\name{gs}
\alias{gs}
\title{Genomic state object for hg19}
\format{A genomic state object for the \code{fullGenome} for hg19
using the XX annotation.}
\description{
The derfinder genomic state object for hg19 created using
TxDb.Hsapiens.UCSC.hg19.knownGene. Only the \code{fullGenome} portion of it
is saved in this package. It takes about 200 seconds to re-make as timed
using \code{system.time()}. See \link[derfinder]{makeGenomicState} for
more information on how to make this type of object.
}
\examples{
## to Amanda: looks like the genomic state object was made using
## Ensembl, unlike the genes info. Ideally they should be from the same
## annotation, so you might want to update the genes info

## system.time(
##     gs <- derfinder::makeGenomicState(
##         txdb = TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene,
##         chrs = paste0('chr', c(1:22, 'X', 'Y'))
##    )$fullGenome
## )
}
\references{
TODO
}
\seealso{
\link{four_panels}
}
\keyword{datasets}
