% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-data.R
\docType{data}
\name{pd}
\alias{pd}
\title{A list of phenotype data.frames}
\format{A list of four data.frames with TODO columns.
\describe{
    \item{Sep }{ phenotype information from TODO,}
    \item{Deg }{ phenotype information from TODO,}
    \item{Cell }{ phenotype information from TODO,}
    \item{Sort }{ phenotype information from TODO.}
}}
\description{
A list of four phenotype data frames used throughput the package.
}
\examples{
##  pd <- list(Sep = pdSep, Deg = pdDeg, Cell = pdCell, Sort = pdSort)

}
\references{
TODO
}
\seealso{
\link{four_panels} \link{brainflowprobes_cov}
}
\keyword{datasets}
