% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-data.R
\docType{data}
\name{genes}
\alias{genes}
\title{GRanges object with the annotated genes for hg19}
\format{A GRanges object with the following annotation columns.
\describe{
    \item{CSS }{ the coding region start,}
    \item{CSE }{ the coding region end,}
    \item{Tx }{ the transcript ID used in TxDb.Hsapiens.UCSC.hg19.knownGene,}
    \item{Geneid }{ the Entrez Gene ID,}
    \item{Gene }{ the Gene symbol,}
    \item{Refseq }{ the RefSeq gene ID,}
    \item{Nexons }{ the number of exons,}
    \item{Exons }{ the exon coordinates.}
}
See \link[bumphunter]{annotateTranscripts} for more information.}
\description{
A GRanges object with the annotated genes for hg19 using
\code{bumphunter::annotateTranscripts} on the
TxDb.Hsapiens.UCSC.hg19.knownGene annotation. Takes about 40 seconds to
re-compute as measured with \code{system.time()}.
}
\examples{

## system.time(
##     genes <- bumphunter::annotateTranscripts(
##         txdb = TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene
##     )
## )

}
\references{
TxDb.Hsapiens.UCSC.hg19.knownGene bumphunter
}
\seealso{
\link{four_panels}
}
\keyword{datasets}
