% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nearest_annotation}
\alias{get_nearest_annotation}
\title{Compute the nearest annotation to the genes in brainflowprobes}
\usage{
get_nearest_annotation(gr, CODING_ONLY = FALSE)
}
\arguments{
\item{gr}{A \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges()}} object.}

\item{CODING_ONLY}{A logical vector of length 1 specifying whether to
subset \link{genes} to only the coding genes. That is, whether to subset
\link{genes} by whether they have a non-NA \code{CSS} value.}
}
\value{
The \code{\link[bumphunter:matchGenes]{bumphunter::matchGenes()}} output for the annotation information
using the \code{genes} object in this package (subset to only the coding elements
if \code{CODING_ONLY} was set to \code{TRUE}).
}
\description{
For a given set of genomic regions, this function computes the nearest
annotation information using the \link{genes} object distributed in this
package.
}
\details{
This is an utility function used by \link{region_info}, \link{four_panels}
and \link{plot_coverage}.
}
\examples{

gr <- GenomicRanges::GRanges('chr10:135379301-135379311:+')

get_nearest_annotation(gr)
get_nearest_annotation(gr, CODING_ONLY = TRUE)

}
\author{
Leonardo Collado-Torres
}
