% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-data.R
\docType{data}
\name{pd}
\alias{pd}
\title{A list of phenotype data.frames}
\format{A list of four data.frames with TODO columns.
\describe{
    \item{Sep }{ phenotype information for samples from
    https://www.biorxiv.org/content/10.1101/567966v1; a data.frame with 23
    rows and 15 columns. Column descriptions: SampleID is the sample name,
    Zone is the RNA fraction of the sample ("Nucleus" or "Cytosol"), Age,
    Sex, and Race list these demographic characteristics, Fetal categorizes
    each sample age as "Fetal" or "Adult", Library categorizes the RNAseq
    library preparation method as polyA selection ("polyA") or rRNA depletion
    ("RiboZero"), RIN_fraction is the RNA Integrity Number for each sample,
    sumMapped is the number of mapped reads, Shortlabels, Label, col, and
    LabelFrac are columns of information for plotting, BigWig is the name of
    the BigWig file for each sample, and files lists the URL for the BigWig
    online.}
    \item{Deg }{ phenotype information for samples from
    https://www.pnas.org/content/114/27/7130; a data.frame with 40 rows and
    16 columns. Column descriptions: DegradationTime is the number of minutes
     the brain tissue for each sample was left on the benchtop at room
     temperature before RNA was extracted, AgeDeath is the age of the donor
     at time of death, Dx lists whether the donor was diagnosed as
     schizophrenic or was a neurotypical control, Sex and Race list these
     demographic characteristics for each sample, pH lists the pH of the
     brain at collection, PMI is the postmortem interval between death and
     brain harvesting, BrNum is the donor number, RIN is the RNA Integrity
     Number for the RNA sample, LibraryProtocol categorizes the RNAseq
     library preparation method as polyA selection ("polyA") or rRNA
     depletion ("Ribo"), SampleID is the ID for the sample, totalAssignedGene
      is the proportion of reads mapping to a gene body, sumMapped is the
      number of mapped reads, BigWig is the name of the BigWig file,
      SampleID_library is the Sample ID and library column values together,
      and files lists the URL for the BigWig online.}
    \item{Cell }{ phenotype information for samples from
    https://www.pnas.org/content/112/23/7285; a data.frame with 466
    rows and 11 columns. Column descriptions: geo_accession is the accession
    number for each sample in the Gene Expression Omnibus, Age is the numeric
     age of each sample, AgeGroup categorizes each sample age as "prenatal"
    or "postnatal", sub_tissue lists whether the sample was from cortex or
    hippocampus, Cluster_color is the color for each sample for plotting,
    Cell_type is the cell identity assigned to each sample, RunName is the
    run name assigned to each sample in the Sequence Read Archive (SRA),
    SubjectID is the subject label, sumMapped is the number of mapped reads,
    BigWig is the name of the BigWig file, and files lists the URL for the
    BigWig online.}
    \item{Sort }{ phenotype information  for samples from
    https://www.biorxiv.org/content/10.1101/428391v2; a data.frame with 12
    rows and 12 columns. Column descriptions: Description categorizes the
    RNAseq library preparation method as polyA selection ("PolyA") or rRNA
    depletion ("Ribo"), SubjectID is the subject number, CellType lists
    whether the sample was labeled by NeuN antibody (neuronal, "NeuN_Plus")
    or not (non-neuronal, "NeuN_Minus"), SampleID combines the sample number,
    cell type and library of each sample in one column, BrNum is the donor
    ID, RIN is the RNA Integrity Number for each RNA sample, Age is the
    numeric age at death, sumMapped is the number of mapped reads, Label and
    col provide information for plotting, BigWig is the name of each BigWig
    file, and files lists the URL for the BigWig online.}
}}
\description{
A list of four phenotype data frames used throughput the package.
}
\examples{
##  pd <- list(Sep = pdSep, Deg = pdDeg, Cell = pdCell, Sort = pdSort)

}
\references{
TODO
}
\seealso{
\link{four_panels} \link{brainflowprobes_cov}
}
\keyword{datasets}
